responsive-dashboard

+++ src/components/Dashboard/AddLinks.jsx
@@ -1,4 +1,4 @@
-import React, { useContext } from "react";
+import React, { useContext, useState } from "react";
 import { setDoc, doc, deleteDoc } from "firebase/firestore";
 import { db } from "firebase-config";
 import { FirebaseContext } from "context/firebase-context";
@@ -6,6 +6,11 @@ import Table from "react-bootstrap/Table";
 import { useForm } from "react-hook-form";
 
 export default function AddLinks() {
+    const [submitState, setSubmitState] = useState({
+        loading: false,
+        success: false,
+        error: false,
+    });
     const {
         register,
         handleSubmit,
@@ -14,19 +19,29 @@ export default function AddLinks() {
 
     const { links } = useContext(FirebaseContext);
     const onSubmit = async (data) => {
+        setSubmitState({ ...submitState, loading: true });
         try {
             await setDoc(doc(db, "Links", data.platform), {
                 platform: data.platform,
                 url: data.url,
             });
         } catch (e) {
-            console.log(e);
+            console.log(e.message);
+            setSubmitState({
+                ...submitState,
+                loading: false,
+                error: e.message,
+            });
         }
+        setSubmitState({ ...submitState, success: true, loading: false });
     };
 
     const onDelete = async (platform) => {
         try {
-            await deleteDoc(doc(db, "Links", platform));
+            const res = window.confirm("Are you sure ?");
+            if (res) {
+                await deleteDoc(doc(db, "Links", platform));
+            }
         } catch (e) {
             console.log(e);
         }
@@ -36,7 +51,6 @@ export default function AddLinks() {
             <Table striped responsive variant='dark' hover>
                 <thead>
                     <tr>
-                        <th></th>
                         <th>Platform</th>
                         <th>URL</th>
                         <th>Delete</th>
@@ -45,13 +59,13 @@ export default function AddLinks() {
                 <tbody>
                     {links?.map((link, index) => (
                         <tr key={index}>
-                            <td>{index + 1}</td>
                             <td>{link.platform}</td>
                             <td>{link.url}</td>
                             <td>
-                                <button onClick={() => onDelete(link.platform)}>
-                                    Delete
-                                </button>
+                                <i
+                                    class='bi bi-trash-fill'
+                                    onClick={() => onDelete(link.platform)}
+                                ></i>
                             </td>
                         </tr>
                     ))}
@@ -59,7 +73,9 @@ export default function AddLinks() {
             </Table>
             <form onSubmit={handleSubmit(onSubmit)}>
                 <fieldset>
-                    <legend>New Link</legend>
+                    <legend>
+                        <h3>New Link</h3>
+                    </legend>
                     <div>
                         <label>Platform{errors.url && <span>*</span>}</label>
                         <select {...register("platform", { required: true })}>
@@ -76,7 +92,24 @@ export default function AddLinks() {
                             {...register("url", { required: true })}
                         />
                     </div>
-                    <button>Submit</button>
+                    {submitState.error && (
+                        <h5 className='alert alert-danger'>Failed!</h5>
+                    )}
+                    {submitState.success && (
+                        <h5 className='alert alert-success'>New Link added.</h5>
+                    )}
+                    {submitState.loading ? (
+                        <button
+                            class='btn btn-outline-primary'
+                            type='button'
+                            disabled
+                        >
+                            <span class='spinner-border spinner-border-sm'></span>
+                            Loading...
+                        </button>
+                    ) : (
+                        <button className='btn btn-primary'>Submit</button>
+                    )}
                 </fieldset>
             </form>
         </div>
